requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
gunicorn>=20.1.0
